refactor synthetic data
