+++ go.mod
@@ -1,6 +1,6 @@
 module babylon/dataloader
 
-go 1.24.4
+go 1.26
 
 require go.mongodb.org/mongo-driver v1.17.4
