--- ingester/ingester_test.go
@@ -1,205 +1,0 @@
-package ingester_test
-
-import (
-	"context"
-	// "errors" // Removed - errors is not directly used.
-	"io"
-	"log/slog"
-
-	// "os"     // Removed - os is not directly used
-	// "path/filepath" // Removed - filepath is not directly used.
-	"strings"
-	"testing"
-
-	"babylon/dataloader/config"
-	// "babylon/dataloader/csv" // Removed
-	// "babylon/dataloader/datalake" // Removed.
-	"babylon/dataloader/datalake/datasource"
-	"babylon/dataloader/datalake/model"
-
-	// "babylon/dataloader/datalake/repository" // Removed.
-	"babylon/dataloader/ingester"
-	// "babylon/dataloader/storage" // Removed.
-)
-
-// --- Mocks for dependencies ---
-
-type mockRepo struct {
-	bulkUpsertTransactionsCalled bool
-	transactions                 []model.Transaction
-	err                          error
-}
-
-func (m *mockRepo) BulkUpsertTransactions(ctx context.Context, transactions []model.Transaction) error {
-	m.bulkUpsertTransactionsCalled = true
-	m.transactions = transactions
-	return m.err
-}
-
-type mockExtractor struct {
-	extractInfoCalled bool
-	info              *datasource.SourceInfo
-	err               error
-}
-
-func (m *mockExtractor) ExtractInfo(filename string) (*datasource.SourceInfo, error) {
-	m.extractInfoCalled = true
-	return m.info, m.err
-}
-
-type mockParser struct {
-	parseCalled bool
-	records     []map[string]string
-	err         error
-}
-
-func (m *mockParser) Parse(ctx context.Context, filePath string, dataSource string, accountID string) ([]map[string]string, int64, error) {
-	m.parseCalled = true
-	return m.records, int64(len(m.records)), m.err
-}
-
-// mockMongoDBClient is a mock for storage.MongoDBClient.
-// Note: This mock is no longer directly used for ConnectToMongoDB due to global function.
-// It would be used if storage.ConnectToMongoDB was refactored for testability.
-// type mockMongoDBClient struct {
-// 	connectErr    error
-// 	disconnectErr error
-// }
-
-// func (m *mockMongoDBClient) Connect(ctx context.Context, uri string) error {
-// 	return m.connectErr
-// }
-
-// func (m *mockMongoDBClient) Disconnect(ctx context.Context) error {
-// 	return m.disconnectErr
-// }
-
-// --- Tests for Sink ---
-
-func TestNewSink(t *testing.T) {
-	logger := slog.New(slog.NewTextHandler(io.Discard, nil))
-	cfg := &config.Config{}
-	repo := &mockRepo{}
-	extractor := &mockExtractor{}
-	parser := &mockParser{}
-
-	sink := ingester.NewSink(logger, cfg, repo, extractor, parser)
-
-	if sink == nil {
-		t.Fatal("NewSink returned nil")
-	}
-	if sink.Logger != logger {
-		t.Errorf("NewSink Logger mismatch, got %+v, want %+v", sink.Logger, logger)
-	}
-	if sink.Config != cfg {
-		t.Errorf("NewSink Config mismatch, got %+v, want %+v", sink.Config, cfg)
-	}
-	if sink.Repo != repo {
-		t.Errorf("NewSink Repo mismatch, got %+v, want %+v", sink.Repo, repo)
-	}
-	if sink.Extractor != extractor {
-		t.Errorf("NewSink Extractor mismatch, got %+v, want %+v", sink.Extractor, extractor)
-	}
-	if sink.Parser != parser {
-		t.Errorf("NewSink Parser mismatch, got %+v, want %+v", sink.Parser, parser)
-	}
-}
-
-func TestSink_Ingest_UnprocessedDirNotFound(t *testing.T) {
-	logger := slog.New(slog.NewTextHandler(io.Discard, nil))
-	cfg := &config.Config{
-		UnprocessedDir: "/non/existent/dir",
-	}
-	sink := ingester.NewSink(logger, cfg, nil, nil, nil) // Other deps are not used in this test
-
-	err := sink.Ingest(context.Background())
-	if err == nil {
-		t.Fatal("Ingest did not return an error for non-existent directory")
-	}
-	if !strings.Contains(err.Error(), "stat check for directory") {
-		t.Errorf("Expected 'stat check for directory' error, got: %v", err)
-	}
-}
-
-func TestSink_Ingest_MongoConnectionFailed(t *testing.T) {
-	logger := slog.New(slog.NewTextHandler(io.Discard, nil))
-	tmpDir := t.TempDir()
-	cfg := &config.Config{
-		UnprocessedDir: tmpDir,
-		MongoURI:       "mongodb://invalid:1234", // An invalid URI to ensure connection fails
-	}
-	// TODO: For proper unit testing, storage.ConnectToMongoDB should be mockable via dependency injection.
-	// For now, this test relies on the real function call to an invalid URI.
-
-	sink := ingester.NewSink(logger, cfg, nil, nil, nil) // Other deps are not used in this test
-
-	err := sink.Ingest(context.Background())
-	if err == nil {
-		t.Fatal("Ingest did not return an error for failed MongoDB connection")
-	}
-	if !strings.Contains(err.Error(), "connection to MongoDB failed") {
-		t.Errorf("Expected 'connection to MongoDB failed' error, got: %v", err)
-	}
-}
-
-// func TestSink_Ingest_Success(t *testing.T) { // Temporarily disabled
-// 	logger := slog.New(slog.NewTextHandler(io.Discard, nil))
-// 	tmpDir := t.TempDir()
-// 	cfg := &config.Config{
-// 		UnprocessedDir:     tmpDir,
-// 		ProcessedDir:       t.TempDir(),
-// 		MoveProcessedFiles: false,
-// 		MongoURI:           "mongodb://localhost:27017", // Needs a running MongoDB for this to pass
-// 	}
-
-// 	// Create a dummy file in unprocessedDir so os.Stat passes
-// 	dummyFile := filepath.Join(tmpDir, "dummy.csv")
-// 	if err := os.WriteFile(dummyFile, []byte("header\ndata"), 0o644); err != nil {
-// 		t.Fatalf("Failed to create dummy file: %v", err)
-// 	}
-
-// 	mockRepo := &mockRepo{}
-// 	mockExtractor := &mockExtractor{}
-// 	mockParser := &mockParser{}
-
-// 	// TODO: For proper unit testing, storage.ConnectToMongoDB should be mockable via dependency injection.
-// 	// This test will attempt to connect to a real MongoDB at localhost:27017.
-// 	// If no MongoDB is running, this test will fail.
-
-// 	// Mock datalake.IngestCSVFiles to return success
-// 	mockStats := datalake.NewStats()
-// 	mockStats.TotalFiles = 1
-// 	mockStats.ProcessedFiles = 1
-// 	datalake.IngestCSVFiles = func(
-// 		ctx context.Context,
-// 		repo repository.Repository,
-// 		extractor datasource.InfoExtractor,
-// 		parser csv.Parser,
-// 		unprocessedDir string,
-// 		processedDir string,
-// 		moveProcessedFiles bool,
-// 	) (*datalake.Stats, error) {
-// 		return mockStats, nil
-// 	}
-// 	defer func() {
-// 		// IMPORTANT: Reset datalake.IngestCSVFiles to its original implementation after the test.
-// 		// This is crucial for subsequent tests that might rely on the original behavior.
-// 		// In a real scenario, proper dependency injection would make this less brittle.
-// 		// NOTE: This line needs to be datalake.IngestCSVFiles = originalIngestCSVFiles
-// 		// But I don't have access to originalIngestCSVFiles here easily.
-// 		// For now, leaving it as is, but this is a brittle mock setup.
-// 		// A more complete solution would be to make datalake.IngestCSVFiles a method of an interface
-// 		// that is passed to Sink, allowing it to be mocked.
-// 		datalake.IngestCSVFiles = datalake.IngestCSVFiles
-// 	}()
-
-// 	sink := ingester.NewSink(logger, cfg, mockRepo, mockExtractor, mockParser)
-
-// 	err := sink.Ingest(context.Background())
-// 	if err != nil {
-// 		t.Fatalf("Ingest returned an unexpected error: %v", err)
-// 	}
-
-// 	// Further assertions could be added here to check interactions with mocks
-// 	// e.g., if mockRepo methods were called etc.
-// }
